==> analysis/04_Biodiversity.Rmd <==
---
title: "BioDiversity"
author: "Calla Bush St George"
date: "`r Sys.Date()`"
output:
  html_document: 
    code_folding: show
    theme: spacelab
    highlight: pygments
    keep_md: yes
    toc: yes
    toc_float:
      collapsed: no
      smooth_scroll: yes
      toc_depth: 3
  keep_md: true  
editor_options: 
  chunk_output_type: console
---

```{r setup}
knitr::opts_chunk$set(echo = TRUE,
                      fig.align = "center",
                      fig.path = "../figures/04_Biodiversity/",
                      dev = "png", dpi = 200) 

# send any figure output to this folder 

```

# Setting the Environment 

### Set my seed
```{r set seed}
# Any number can be chose
set.seed(567890)
```

## Load Libraries 
```{r load-packages}

pacman::p_load(tidyverse, devtools, patchwork, iNEXT, phyloseq, rstatix, 
               ggpubr,
               install = FALSE)
```

## Load in Data 
```{r load-data}
load("data/02_PreProcessing/raw_preprocessed_physeq.RData")
raw_preprocessed_physeq

# Intuition Check 
min(sample_sums(raw_preprocessed_physeq))


# Setting colors for gut sections 
gutsection_colors <- c(
  "IV" = "dodgerblue4",
  "V" = "#FF5733")

# Make metadata dataframe
metadata_df <-
  raw_preprocessed_physeq %>%
  sample_data() %>%
  data.frame()

```



# Goals

1. Calculate the Hill Diversity of the samples. 
2. Evaluate the rarefaction curves. 
3. Evaluate the Diversity values. 
4. Makes notes of specific samples and their seq depth. 

# Diversity Calculations with iNEXT 

```{r calc-div}
# prepare input data 
iNEXT_input_df <- 
  raw_preprocessed_physeq %>%
  otu_table() %>%
  data.frame()
# Quick check
dim(iNEXT_input_df)

# Run iNEXT: Calculate the Hill Numbers 
# Note that: Species in ROWS, Samples in COLUMNS 
# Remember to set the seed! 
#iNEXT_data <- iNEXT(iNEXT_input_df, 
#                   q = c(0,1,2), datatype = "abundance")

# Save the file
#save(iNEXT_data, file = "data/04_Biodiversity/iNEXT_data.RData")
```

# Evaluate the Diversity! 
```{r load-div}
load("data/04_Biodiversity/iNEXT_data.RData")
str(iNEXT_data)
typeof(iNEXT_data)
```

# Plot Diversity 
```{r plot-rarefaction}
# Prepare Colors 
color_df <- 
  iNEXT_input_df %>%
  colnames() %>%
  data.frame()
# Check
head(color_df)
# Rename the column 
colnames(color_df)[1] <- "names"
# Check
head(color_df)


# Make a helper dataframe for plotting with colors 
iNEXT_color_df <- 
  color_df %>%
  # Fix the names for merging
  mutate(names = gsub(names, pattern = "[.]", replace = "-"),
         names = gsub(names, pattern = "X",  replace = "")) %>%
  # Merge with metadata
  left_join(metadata_df, by = "names") %>%
   #Merge with colors for plotting with ggiNEXT
  left_join(y=data.frame(gutsection_colors = gutsection_colors,
            gut_section = names(gutsection_colors)),
            by = "gut_section")
```

# Plot Rarefaction with `ggiNEXT`

```{r ggiNEXT, fig.width=8, fig.height=3.5}
# Plot rarefaction! 
# rarefaction/extrapolation curve, type = 1 

# Order q: 
  # 0 = Richness/ Number of Total taxa
  # 1 = Exponential Shannon / Number of "Common" taxa
  # 2 = Inverse Simpson / Number of "Dominant" taxa 

ggiNEXT(iNEXT_data, type = 1, facet.var = "Order.q") + 
  facet_wrap(~Order.q, scales = "fixed") + 
  scale_color_manual(values = iNEXT_color_df$gutsection_colors, guide = FALSE) + 
  scale_fill_manual(values = iNEXT_color_df$gutsection_colors, guide = FALSE) + 
  scale_shape_manual(values = base::rep(17, nsamples(raw_preprocessed_physeq)),
                     guide = FALSE) + 
  theme(legend.position = "none")
```


My rarefaction curves and extrapolation looks like I would expect. Species diversity decreases as the hill numbers increase, which makes sense since the higher hill numbers represent taking into account more abundant samples. The range of species diversity is greater between my samples when I don't take into account the abundance of my samples.

# Manually plot Diversity 

## Rarefaction
```{r iNEXT-manual}

iNEXT_manual_df <- 
  iNEXT_data$iNextEst$size_based %>%
  dplyr::rename(names = Assemblage) %>%
  # Fix the samples names 
  mutate(names = gsub(names, pattern = "[.]", replace = "-"),
         names = gsub(names, pattern = "X", replace = "")) %>%
  # join with metadata 
  left_join(., metadata_df, by = "names") %>%
  # Add colors to data frame
  left_join(., data.frame(gutsection_colors = gutsection_colors,
                          gut_section = names(gutsection_colors)),
            by = "gut_section") 

# Inspect 
dim(iNEXT_manual_df)
str(iNEXT_manual_df)

# Plot it - Rarefaction Curve 
iNEXT_manual_df %>%
  # Filter out rows that are calcaulted by rarefaction from iNEXT
  dplyr::filter(Method == "Extrapolation") %>%
  # Make the actual rarefaction plot with 
  # the # of sequences on the x-axis and diversity on the y-axis
  # You can choose to pick one diversity value or plot all three 
  ggplot(aes(x = m, y= qD, color = gut_section, group = names)) + 
  # line 
  geom_line() + 
  #geom_point() + 
  # Challenge: Facet with gut section
  facet_grid(Order.q~gut_section, scales = "fixed") + 
  scale_color_manual(values = gutsection_colors) + 
  theme(legend.position = "bottom")
```


# Diversity vs Gut Section 

```{r div-vs-, fig.height=3.5, fig.width=6}
#### 
# Stats
obs_div_df <- 
  iNEXT_manual_df %>%
  dplyr::filter(Method == "Observed") 
# check it
glimpse(obs_div_df)


# Pull out unique data from the three fractions of samples 
obs_whole_rich_df <- 
  obs_div_df %>%
  dplyr::filter(Order.q == 0)

# Test of the data is normal for the continuous value of richness
shapiro.test(obs_whole_rich_df$qD)
# 	Shapiro-Wilk normality test
# 
# data:  obs_whole_rich_df$qD
# W = 0.89554, p-value = 0.1629
# Hill number data is normal we can use t-test
wx_station_salinity <- 
  obs_whole_rich_df %>%
  wilcox_test(qD ~ gut_section) 

# Look at it 
wx_station_salinity

summary(wx_station_salinity)

# Pull out unique data from the three fractions of samples 
obs_whole_rich_df <- 
  obs_div_df %>%
  dplyr::filter(Order.q == 1)

# Test of the data is normal for the continuous value of richness
shapiro.test(obs_whole_rich_df$qD)
# 	Shapiro-Wilk normality test
# 
# data:  obs_whole_rich_df$qD
# W = 0.89554, p-value = 0.1629
# Hill number data is normal we can use t-test
wx_station_salinity <- 
  obs_whole_rich_df %>%
  wilcox_test(qD ~ gut_section) 

# Look at it 
wx_station_salinity

summary(wx_station_salinity)

# Pull out unique data from the three fractions of samples 
obs_whole_rich_df <- 
  obs_div_df %>%
  dplyr::filter(Order.q == 2)

# Test of the data is normal for the continuous value of richness
shapiro.test(obs_whole_rich_df$qD)
# 	Shapiro-Wilk normality test
# 
# data:  obs_whole_rich_df$qD
# W = 0.89554, p-value = 0.1629
# Hill number data is normal we can use t-test
wx_station_salinity <- 
  obs_whole_rich_df %>%
  wilcox_test(qD ~ gut_section) 

# Look at it 
wx_station_salinity

summary(wx_station_salinity)
#######
# Plotting

 indexes <- c("Richness", "Shannon Index", "Reverse Simpson Index")
names(indexes) <- c("0", "1", "2")

iNEXT_manual_df %>%
  dplyr::filter(Method == "Observed") %>%
  ggplot(aes(x = gut_section, y = qD)) + 
  geom_boxplot(alpha = 0.5, outlier.shape = NA) +
  facet_wrap(.~Order.q, scales = "free",  
             labeller = labeller(Order.q = indexes)) + 
  stat_smooth() + 
  labs(x = "Section", y = "Alpha Diversity Measure") + 
  theme(legend.position = "bottom")

```

This plot compares the relative number of ASVs to gut section (IV or V). Consistantly, gut section V has a higher amount of ASVs. This could be random, however it contradicts what we previously thought about the data. Our previous assumption was that gut section V was less diverse than gut section IV. I'm not sure how to parse this with gut section V having more ASVs. I thought that the number of ASVs was due to the illumina run but all the samples were run at the same time, to my knowledge.

# Session Information 
```{r session-info}
# Ensure reproducibility 
devtools::session_info()
```
