==> analysis/05_Community_Analysis.Rmd <==
---
title: "Between sample/beta diversity of Microbes by gut section"
author: "Calla Bush St George"
date: "`r Sys.Date()`"
output:
  html_document: 
    code_folding: show
    theme: spacelab
    highlight: pygments
    keep_md: no
    toc: yes
    toc_float:
      collapsed: no
      smooth_scroll: yes
      toc_depth: 3
  keep_md: true  
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      fig.align = "center",
                      fig.path = "../figures/05_Community_Analysis/")
```

# Goals

1. Load in phyloseq data with rooted tree.
2. Evaluate sequencing depth and remove sample.
3. Normalize the read counts between samples.
4. Calculate community dissimilarities. Numbers between 0 and 1. If 0, completely similar versus if they are 1, then they are completely dissimilar. 
    a. *Sorensen*: Presence/ Absence. Weighted by number of shared taxa. Shared species as a binary-valye. Abundance-unweighted. 
    b. *Bray-Curtis*: Relative abundance. Weighted by number of shared taxa. Shared abundant species: abundance weighted. 
    c. *(Abundance) Weighted UNIFRAC*: Consider abundant species and where they fall on the tree. 
5. Visualize the community data with two unconstricted ordinations:
    a. *PCoA*: Linear method. Uses matrix algebra to calculate eigenvalye. Calculate how much variation is explained by each axis.  Can choose to view axis 1, 2, 3, etc. and plot them together. 
    b. *NMDS*: Non-linear method. Collapse multiple axes into two (or three) dimensions. Can see more axes of variation into fewer axes. Always need to report a stress value. (Ideally less than 0.15)
6. Run statistics with PERMANOVA and betadispR.

# Setup

## Load Libraries
```{r load-packages}
# Install packages if needed

pacman::p_load(tidyverse, devtools, phyloseq, patchwork, vegan, ggpubr, rstatix,
               ggplot2,install = FALSE)
```

## Load in colors
```{r load-colors}
gut_section_colors <- c(
  "IV" = "dodgerblue4",
  "V" = "#FF5733")
```


## Load in data 
```{r load-physeq}
# Load in rooted phylogenetic tree!
load("data/03_Phylogenetic_Tree/phytree_preprocessed_physeq.RData")
unrooted_physeq
midroot_physeq
```

# Explore Read Counts

## Raw Read Depth
```{r calculate-seq-depth}
# Sequence depth will inform us on how we want to normalize our data
# Calculate the total number of reads per sample
raw_total_seqs_df <-
  unrooted_physeq %>%
  # Calculate the sample read sums
  sample_sums() %>%
  data.frame()

# Name the column
colnames(raw_total_seqs_df)[1] <- "TotalSeqs"

head(raw_total_seqs_df)

# Make a histogram of raw reads
raw_seqs_histogram <-
  raw_total_seqs_df %>%
  ggplot(aes(x = TotalSeqs)) +
  geom_histogram(bins = 50) +
  scale_x_continuous(limits = c(0, 50000)) +
  labs(title = "Raw Sequencing Depth Distribution") + 
  theme_bw()

```

## Remove lowly sequenced sample

## Normalize Read Counts
```{r scale-reads}

### scale_reads function and matround function
#################################################################################### 
# Function to scale reads: http://deneflab.github.io/MicrobeMiseq/ 
# Scales reads by 
# 1) taking proportions
# 2) multiplying by a given library size of n
# 3) rounding 
# Default for n is the minimum sample size in your library
# Default for round is floor

matround <- function(x){trunc(x+0.5)}

scale_reads <- function(physeq, n = min(sample_sums(physeq)), round = "round") {
  
  # transform counts to n
  physeq.scale <- transform_sample_counts(physeq, function(x) {(n * x/sum(x))})
  
  # Pick the rounding functions
  if (round == "floor"){
    otu_table(physeq.scale) <- floor(otu_table(physeq.scale))
  } else if (round == "round"){
    otu_table(physeq.scale) <- round(otu_table(physeq.scale))
  } else if (round == "matround"){
    otu_table(physeq.scale) <- matround(otu_table(physeq.scale))
  }
  
  # Prune taxa and return new phyloseq object
  physeq.scale <- prune_taxa(taxa_sums(physeq.scale) > 0, physeq.scale)
  return(physeq.scale)
  
  }

```

Rescale all reads so they all represent the count of the lowest number of sequence reads.
We will expect each sample to have # of reads around 2200

This is where one might decide to use rarefaction to normalize the data. 

## Scale reads and check the distribution of the seq depth
```{r scale-physeq}
min(sample_sums(midroot_physeq))

# Scale reads by the above function
scaled_rooted_physeq <-
  midroot_physeq %>%
  scale_reads(round = "matround")

# Calculate read depth
## Look at total number of sequences in each sample and compare to what we had before

scaled_total_seqs_df <- 
  scaled_rooted_physeq %>%
  sample_sums() %>%
  data.frame()

head(scaled_total_seqs_df)

# Change first column name to be "TotalSeqs"
colnames(scaled_total_seqs_df)[1] <- "TotalSeqs"

# Inspect
head(scaled_total_seqs_df)

# Check range of data
min_seqs <-
  min(scaled_total_seqs_df)
max_seqs <-
 max(scaled_total_seqs_df)
# Range of seqs
max_seqs - min_seqs

# Plot histogram
scaled_total_seqs_df %>%
  ggplot(aes(x = TotalSeqs)) +
  geom_histogram(bins = 50) +
  scale_x_continuous(limits = c(0, 10000)) +
  labs(title = "Scaled Sequencing Depth at 7131") + 
  theme_bw()

head(scaled_total_seqs_df)

```

# Calculate & Visualize Community Dissimiliarity

Exploratory analyses from the Paily & Shankar (2016) paper, which is using unconstrained ordination methods like PCoA. 

## Sorenson PCoA
```{r sorenson-pcoa}
# Calculate sorenson dissimularity: Abundance-unweighted of shared taxa
scaled_soren_pcoa <-  
  ordinate(
    physeq = scaled_rooted_physeq,
    method = "PCoA",
    distance = "bray", binary = TRUE)

#str(scaled_soren_pcoa)

# Plot the ordination 
soren_gut_section_pcoa <- plot_ordination(
  physeq = scaled_rooted_physeq,
  ordination = scaled_soren_pcoa,
  color = "gut_section",
  shape = "gut_section",
  title = "Sorensen PCoA") +
  geom_point(size=5, alpha = 0.5, aes(color = gut_section)) +
  scale_shape_manual(values = c(15,16)) +
  scale_color_manual(values = gut_section_colors) + 
  theme_bw()
# Show the plot 
soren_gut_section_pcoa
```

Here, we are evaluating the shared taxa by presence/absence (abundance-unweighted) in the Sorensen metric. 

Note that:  

- Axis 1 = ~31% of variation  
- Axis 2 = ~14% of variation 

This means we explain 45% of the variation in the data in these two axes. 



## Bray-Curtis PCoA
```{r bray-PCoA}
# Calculate the BC distance
scaled_BC_pcoa <- 
  ordinate(
    physeq = scaled_rooted_physeq,
    method = "PCoA",
    distance = "bray")

# Plot the PCoA
bray_gut_section_pcoa <- 
  plot_ordination(
    physeq = scaled_rooted_physeq,
    ordination = scaled_BC_pcoa,
    color = "gut_section",
    shape = "gut_section",
    title = "Bray-Curtis PCoA") +
  geom_point(size=5, alpha = 0.5, aes(color = gut_section)) +
  scale_shape_manual(values = c(15,16)) +
  scale_color_manual(values = c(gut_section_colors)) + 
  theme_bw()
bray_gut_section_pcoa
```

Here, we are evaluating the shared taxa and then weighting them by their abundances, which provides more influence for species that are more abundant. 

Note that:  

- Axis 1 = ~30% of variation  
- Axis 2 = ~14% of variation 

This means we explain 44% of the variation in the data in these two axes, which is almost exactly the same as the previous plot with the Sorensen Dissimilarity. 
Abundance does NOT seem to have an influence!!

## Weighted-Unifrac PCoA
```{r wUnifrac-PCoA}
# Calculate the BC distance
scaled_wUNI_pcoa <- 
  ordinate(
    physeq = scaled_rooted_physeq,
    method = "PCoA",
    distance = "wunifrac")

# Plot the PCoA
wUNI_gut_section_pcoa <- 
  plot_ordination(
    physeq = scaled_rooted_physeq,
    ordination = scaled_wUNI_pcoa,
    color = "gut_section",
    shape = "gut_section",
    title = "Weighted Unifrac PCoA") +
  geom_point(size=5, alpha = 0.5, aes(color = gut_section)) +
  scale_shape_manual(values = c(15,16,17,18, 19, 20)) +
  scale_color_manual(values = c(gut_section_colors)) + 
  theme_bw()
wUNI_gut_section_pcoa
```

Here, we are evaluating the shared taxa and then weighting them by their abundances, which provides more influence for species that are more abundant. 


Note that:   

- Axis 1 = ~60% of variation  
- Axis 2 = ~22% of variation 

This means we explain *82*% of the variation in the data in these two axes (!!!), which is significantly more than the previous plots with the taxonomic dissimilarity measures. (Almost double.) Here, phylogeny seems to be very important! This means that taxa that are abundant are found in different places in the phylogenetic tree. Therefore, the evoultionary distances (aka the branch lengths) and their abundances seem to have a major influence!!

## Combine PCoAs

Let's plot all three together into one plot to have a concise visualization of the three metrics. 
```{r pcoa-together, fig.width=8, fig.height=3.5}
(soren_gut_section_pcoa + theme(legend.position = "none")) + 
  (bray_gut_section_pcoa + theme(legend.position = "none")) + 
    (wUNI_gut_section_pcoa + theme(legend.position = "none"))
```

## NMDS 

### Weighted Unifrac

Since we did 3 of the dissimilarity metrics for the PCoA, let's just plot one example of them for the NMDS plotting. Here, we will use weighted Unifrac 
```{r wUnifrac-NMDS}
# Calculate the Weighted Unifrac distance
scaled_wUNI_nmds <- 
  ordinate(
    physeq = scaled_rooted_physeq,
    method = "NMDS",
    distance = "wunifrac")

# Plot the PCoA
wUNI_gut_section_nmds <- 
  plot_ordination(
    physeq = scaled_rooted_physeq,
    ordination = scaled_wUNI_nmds,
    color = "gut_section") +
  geom_point(size=5, alpha = 0.5, aes(color = gut_section)) +
  scale_shape_manual(values = c(15,16)) +
  scale_color_manual(values = c(gut_section_colors)) + 
  theme_bw() + labs(color = "Gut Section")
wUNI_gut_section_nmds
```

We can see from above the plot that the stress value is ~0.02, which is very acceptable. And, It seems important to emphasize that the PCoA and the NMDS plot both look pretty similar! 

In this case, I would always prefer to report the PCoA results because they are linear and provide a lot more post-hoc analyses to follow up with. In addition, it's helpful to only have 2 axes of variation and show how much variation is explained. 

```{r combined-wUnifrac, fig.width=6, fig.height=3.5}
(wUNI_gut_section_pcoa + theme(legend.position = "none")) + 
  (wUNI_gut_section_nmds + theme(legend.position = "none"))
```


# Statistical Significance Testing 

## PERMANOVA 
```{r PERMANOVA}
# Calculate all three of the distance matrices
scaled_sorensen_dist <- phyloseq::distance(scaled_rooted_physeq, method = "bray", binary = TRUE)
scaled_bray_dist <- phyloseq::distance(scaled_rooted_physeq, method = "bray")
scaled_wUnifrac_dist <- phyloseq::distance(scaled_rooted_physeq, method = "wunifrac")

# make a data frame from the sample_data
# All distance matrices will be the same metadata because they 
# originate from the same phyloseq object. 
metadata <- data.frame(sample_data(scaled_rooted_physeq))

# Adonis test
# In this example we are testing the hypothesis that the five stations
# that were collected have different centroids in the ordination space 
# for each of the dissimilarity metrics, we are using a discrete variable 
adonis2(scaled_sorensen_dist ~ gut_section, data = metadata)
adonis2(scaled_bray_dist ~ gut_section, data = metadata)
adonis2(scaled_wUnifrac_dist ~ gut_section, data = metadata)
```

Note that:  

- R2 = the percent variation explained. 
- F = the F-Statistic, which represents the importance value. 
- Pr(>F) = the pvalue 


Above, we see that the most variation is explained by the weighted unifrac, which explains ~55% of the variation in the data and also has the highest F-statistic.



```{r PERMANOVA-multiVariable}
# We might also care about other variables
# Here, we will add date and fraction as variables
# multiplicative model ORDER MATTERS! 
adonis2(scaled_sorensen_dist ~ gut_section * year, data = metadata)
adonis2(scaled_bray_dist ~ gut_section * year, data = metadata)
# Note that the ORDER MATTERS!
adonis2(scaled_wUnifrac_dist ~ gut_section * year, data = metadata)
adonis2(scaled_wUnifrac_dist ~ year * gut_section, data = metadata)
```

We can also run tests that include additive (+) or multipliciatve models, which include the interaction term between variables. 

## BetaDispR

The PERMANOVA is sensitive to variance/dispersion in the data. Therefore, we need to run a homogeneity of dispersion test to test for the sensitivity of our PERMANOVA results to variance. 
```{r betadispR}
# Homogeneity of Disperson test with beta dispr
# Sorensen 
beta_soren_gut_section <- betadisper(scaled_sorensen_dist, metadata$gut_section)
permutest(beta_soren_gut_section)

# Bray-curtis 
beta_bray_gut_section <- betadisper(scaled_bray_dist, metadata$gut_section)
permutest(beta_bray_gut_section)

# Weighted Unifrac 
beta_bray_gut_section <- betadisper(scaled_wUnifrac_dist, metadata$gut_section)
permutest(beta_bray_gut_section)
```

Above, our variance is not impacted by gut section. 

# Taxonomic Composition

## Phylum
```{r phylum-colors}
# Set the phylum colors
phylum_colors <- c(
  Acidobacteriota = "navy", 
  Actinobacteriota = "darkslategray2", 
  Armatimonadota = "deeppink1",
  Alphaproteobacteria = "plum2", 
  Bacteroidota = "gold", 
  Betaproteobacteria = "plum1", 
  Bdellovibrionota = "red1",
  Chloroflexi="black", 
  Crenarchaeota = "firebrick",
  Cyanobacteria = "limegreen",
  Deltaproteobacteria = "grey", 
  Desulfobacterota="magenta",
  Fibrobacterota = "darkgreen",
  Bacillota = "#3E9B96",
  Gammaproteobacteria = "greenyellow",
  "Marinimicrobia (SAR406 clade)" = "yellow",
  Myxococcota = "#B5D6AA",
  Nitrospirota = "palevioletred1",
  Pseudomonadota = "royalblue",
  Planctomycetota = "darkorange", 
  Spirochaetota = "olivedrab",
  Thermoplasmatota = "green",
  Verrucomicrobiota = "darkorchid1")
```

#### Plot Phylum Composition
```{r phylum-composition}
# Goal is to calculate the phylum relative abundance
# Note: the read depth must be normalized in some way: scaled_reads
level_order <- 
  c('568_4','E1_4','E2_4','E3_4','568_5', '581_5','611_5', 'E05_5', 'E1_5',
    'E2_5', 'E3_5') 

phylum_df <-
  scaled_rooted_physeq %>%
  # agglomerate at the phylum level
  tax_glom(taxrank = "Phylum") %>%
  # Transform counts to relative abundance
  transform_sample_counts(function (x) {x/sum(x)}) %>%
  # Melt to a long format
  psmelt() %>%
  # Filter out phyla that are less than one percent - get rid of low abundant Phyla
  dplyr::filter(Abundance > 0.01)
  # fix the order of date
  
# Stacked bar plot with all Phyla
# Plot Phylum Abundances - make sure to load phylum_colors
phylum_df %>%
  # Warning: Its important to have one sample per x value,
  # Otherwise, it will take the sum between multiple samples
  ggplot(aes(x = factor(names, level = level_order),
             y = Abundance, fill = Phylum)) +
  facet_grid(.~gut_section) + 
  geom_bar(stat = "identity", color = "black") +
  labs(title = "Gut Section Phylum Composition") +
  scale_fill_manual(values = phylum_colors) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1))


# Plot Phylum Abundances - make sure to load phylum_colors
phylum_df %>%
  # Warning: Its important to have one sample per x value,
  # Otherwise, it will take the sum between multiple samples
  ggplot(aes(x = factor(names, level = level_order),
             y = Abundance, fill = Phylum)) +
  geom_bar(stat = "identity", color = "black") +
  labs(title = "Gut Section Phylum Composition") + 
  scale_fill_manual(values = phylum_colors) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1)) +
  labs(y ="Relative Abundance",x = "Sample")

# Make each Phyla its own row
#fixed scale
phylum_df %>%
  ggplot(aes(x = names, y = Abundance, fill = Phylum)) +
  facet_grid(Phylum~gut_section) + 
  geom_bar(stat = "identity", color = "black") +
  labs(title = "Gut Section Phylum Composition") + 
  scale_fill_manual(values = phylum_colors) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1))

# free scale
phylum_df %>%
  ggplot(aes(x = names, y = Abundance, fill = Phylum)) +
  facet_grid(Phylum~gut_section, scale = "free") + 
  geom_bar(stat = "identity", color = "black") +
  labs(title = "Gut Section Phylum Composition") + 
  scale_fill_manual(values = phylum_colors) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1))
  
# Narrow in on a specific group

# Bacteroidota - y: abundance, x: gut section, dot plot + boxplot
Bacteroidota_phylum <-
  phylum_df %>%
  dplyr::filter(Phylum == "Bacteroidota") %>%
  # build the plot
  ggplot(aes(x = gut_section, y = Abundance, fill = gut_section, color = gut_section)) + 
  geom_boxplot(alpha = 0.5, outlier.shape = NA) +
  # outliers not plotted here in boxplot
  geom_jitter() +
  theme_bw() +
  labs(title = "Bacteroidota Phylum Abundance") +
  scale_color_manual(values = gut_section_colors) +
  scale_fill_manual(values = gut_section_colors) +
  stat_compare_means(paired = FALSE)
Bacteroidota_phylum

# Cyanobacteria - y: abundance, x: gut section, dot plot + boxplot
Cyanobacteria_phylum <-
  phylum_df %>%
  dplyr::filter(Phylum == "Cyanobacteria") %>%
  # build the plot
  ggplot(aes(x = gut_section, y = Abundance, fill = gut_section, color = gut_section)) + 
  geom_boxplot(alpha = 0.5, outlier.shape = NA) +
  # outliers not plotted here in boxplot
  geom_jitter() +
  theme_bw() +
  labs(title = "Cyanobacteria Phylum Abundance") +
  scale_color_manual(values = gut_section_colors) +
  scale_fill_manual(values = gut_section_colors) +
  stat_compare_means(paired = FALSE)
Cyanobacteria_phylum

# Desulfobacterota - y: abundance, x: gut section, dot plot + boxplot
Desulfobacterota_phylum <-
  phylum_df %>%
  dplyr::filter(Phylum == "Desulfobacterota") %>%
  # build the plot
  ggplot(aes(x = gut_section, y = Abundance, fill = gut_section, color = gut_section)) + 
  geom_boxplot(alpha = 0.5, outlier.shape = NA) +
  # outliers not plotted here in boxplot
  geom_jitter() +
  theme_bw() +
  labs(title = "Desulfobacterota Phylum Abundance") +
  scale_color_manual(values = gut_section_colors) +
  scale_fill_manual(values = gut_section_colors) +
  stat_compare_means()
Desulfobacterota_phylum

# Fibrobacterota - y: abundance, x: gut section, dot plot + boxplot
Fibrobacterota_phylum <-
  phylum_df %>%
  dplyr::filter(Phylum == "Fibrobacterota") %>%
  # build the plot
  ggplot(aes(x = gut_section, y = Abundance, fill = gut_section, color = gut_section)) + 
  geom_boxplot(alpha = 0.5, outlier.shape = NA) +
  # outliers not plotted here in boxplot
  geom_jitter() +
  theme_bw() +
  labs(title = "Fibrobacterota Phylum Abundance") +
  scale_color_manual(values = gut_section_colors) +
  scale_fill_manual(values = gut_section_colors) +
  stat_compare_means()
Fibrobacterota_phylum

# Firmicutes - y: abundance, x: gut section, dot plot + boxplot
Bacillota_phylum <-
  phylum_df %>%
  dplyr::filter(Phylum == "Bacillota") %>%
  # build the plot
  ggplot(aes(x = gut_section, y = Abundance, fill = gut_section, color = gut_section)) + 
  geom_boxplot(alpha = 0.5, outlier.shape = NA) +
  # outliers not plotted here in boxplot
  geom_jitter() +
  theme_bw() +
  labs(title = "Bacillota Phylum Abundance") +
  scale_color_manual(values = gut_section_colors) +
  scale_fill_manual(values = gut_section_colors) +
  stat_compare_means()
Bacillota_phylum

# Proteobacteria - y: abundance, x: gut section, dot plot + boxplot
Pseudomonadota_phylum <-
  phylum_df %>%
  dplyr::filter(Phylum == "Pseudomonadota") %>%
  # build the plot
  ggplot(aes(x = gut_section, y = Abundance, fill = gut_section, color = gut_section)) + 
  geom_boxplot(alpha = 0.5, outlier.shape = NA) +
  # outliers not plotted here in boxplot
  geom_jitter() +
  theme_bw() +
  labs(title = "Pseudomonadota Phylum Abundance") +
  scale_color_manual(values = gut_section_colors) +
  scale_fill_manual(values = gut_section_colors) +
  stat_compare_means()
Pseudomonadota_phylum

# Spirochaetota- y: abundance, x: gut section, dot plot + boxplot
Spirochaetota_phylum <-
  phylum_df %>%
  dplyr::filter(Phylum == "Spirochaetota") %>%
  # build the plot
  ggplot(aes(x = gut_section, y = Abundance, fill = gut_section, color = gut_section)) + 
  geom_boxplot(alpha = 0.5, outlier.shape = NA) +
  # outliers not plotted here in boxplot
  geom_jitter() +
  theme_bw() +
  labs(title = "Spirochaetota Phylum Abundance") +
  scale_color_manual(values = gut_section_colors) +
  scale_fill_manual(values = gut_section_colors)+
  stat_compare_means()
Spirochaetota_phylum

# Verrucomicrobiota - y: abundance, x: gut section, dot plot + boxplot
Verrucomicrobiota_phylum <-
  phylum_df %>%
  dplyr::filter(Phylum == "Verrucomicrobiota") %>%
  # build the plot
  ggplot(aes(x = gut_section, y = Abundance, fill = gut_section, color = gut_section)) + 
  geom_boxplot(alpha = 0.5, outlier.shape = NA) +
  # outliers not plotted here in boxplot
  geom_jitter() +
  theme_bw() +
  labs(title = "Verrucomicrobiota Phylum Abundance") +
  scale_color_manual(values = gut_section_colors) +
  scale_fill_manual(values = gut_section_colors) +
  stat_compare_means()
Verrucomicrobiota_phylum

```


## Order
```{r order-colors}
# Set Order colors
order_colors <- c(
  Bacteroidales = "darkslategray2", 
  Christensenellales = "gold",
  "Clostridia vadinBB60 group" = "deeppink1",
  Clostridiales = "limegreen",  
  Cytophagales = "gray",
  Desulfobacterales = "magenta4", 
  Desulfovibrionales = "black",
  Enterobacterales = "firebrick",
  Fibrobacterales = "yellow",
  Gastranaerophilales = "greenyellow",
  Lachnospirales = "yellow4", 
  Opitutales ="magenta",
  Oscillospirales = "palevioletred1",
  "P.palmC41"  = "orange3",
  "Peptostreptococcales-Tissierellales" = "lightgreen",
  Pseudomonadales = "purple2",
  Spirochaetales = "royalblue",
  Verrucomicrobiales  = "plum2",
  Victivallales = "darkgray",
  "WCHB1-41"=  "violetred"
)
```

```{r order-composition}
Order_df <-
  scaled_rooted_physeq %>%
  # agglomerate at the phylum level
  tax_glom(taxrank = "Order") %>%
  # Transform counts to relative abundance
  transform_sample_counts(function (x) {x/sum(x)}) %>%
  # Melt to a long format
  psmelt() %>%
  # Filter out phyla that are less than one percent - get rid of low abundant Phyla
  dplyr::filter(Abundance > 0.01)

#section_IV <-
#  c('568_4','E1_4','E2_4','E3_4')

#section_IV_df <- filter(Order_df, names == c('568_4','E1_4','E2_4','E3_4') )

#section_V <-
#  c('568_5', '581_5','611_5', 'E05_5', 'E1_5', 'E2_5', 'E3_5')

#section_V_df <- filter(Order_df, names == section_V )

#save(Order_df, file = "data/05_community_analysis/ASV_counts_order.RData")

#write.csv2(Order_df, file = "data/05_community_analysis/ASV_counts_order.csv")

#write.table(Order_df, "data/05_community_analysis/ASV_counts_order.tsv",
#            sep = "\t", quote = FALSE, col.names = NA)

# Facet by gut section
Order_df %>%
  # Warning: Its important to have one sample per x value,
  # Otherwise, it will take the sum between multiple samples
  ggplot(aes(x = names, y = Abundance, fill = Order)) +
  facet_grid(.~gut_section) + 
  geom_bar(stat = "identity", color = "black") +
  #scale_fill_manual(values = order_colors) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1))

level_order <- 
  c('568_4','E1_4','E2_4','E3_4','568_5', '581_5','611_5', 'E05_5', 'E1_5', 'E2_5', 'E3_5') 

Order_df %>%
  # Warning: Its important to have one sample per x value,
  # Otherwise, it will take the sum between multiple samples
  ggplot(aes(x = factor(names, level = level_order), 
             y = Abundance, fill = Order)) +
  #facet_grid(.~gut_section) + 
  geom_bar(stat = "identity", color = "black") +
  scale_fill_manual(values = order_colors) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1)) +
  labs(y ="Relative Abundance",x = "Sample")

# Section IV

section_IV_df <- read.table("data/05_community_analysis/section_IV.tsv", sep = ",", skip = 1)

head(section_IV_df)

colnames(section_IV_df) <- c("", "OTU", "Sample", "Abundance", "names", 
                             "host_species", "gut_section", "region","loction", "year",
                             "month", "day", "sample_lab", "Time", "SL.mm.", "FL.mm.",
                             "TW..g.", "GW..g.", "Sex", "Sample_or_Control", 
                             "input", "filtered", "denoisedF", "denoisedR", 
                             "merged", "nochim", "perc_reads_retained", "Kingdom",
                             "Phylum", "Class", "Order")
head(section_IV_df)

section_IV_df %>%
  ggplot(aes(x = names, 
             y = Abundance, fill = Order)) +
  geom_bar(stat = "identity", color = "black") +
  scale_fill_manual(values = order_colors) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1)) +
  labs(y ="Relative Abundance",x = "Sample")

# Section V

section_V_df <- read.table("data/05_community_analysis/section_V.tsv", sep = ",", skip = 1)

head(section_V_df)

colnames(section_V_df) <- c("", "OTU", "Sample", "Abundance", "names", 
                             "host_species", "gut_section", "region","loction", "year",
                             "month", "day", "sample_lab", "Time", "SL.mm.", "FL.mm.",
                             "TW..g.", "GW..g.", "Sex", "Sample_or_Control", 
                             "input", "filtered", "denoisedF", "denoisedR", 
                             "merged", "nochim", "perc_reads_retained", "Kingdom",
                             "Phylum", "Class", "Order")
head(section_V_df)

section_V_df %>%
  ggplot(aes(x = names, 
             y = Abundance, fill = Order)) +
  geom_bar(stat = "identity", color = "black") +
  scale_fill_manual(values = order_colors) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1)) +
  labs(y ="Relative Abundance",x = "Sample")
```

## Genus

```{r genus-colors}
# Set the phylum colors
Genus_colors <- c(
  Akkermansia = "navy",  
  "Christensenellaceae R-7 group" = "greenyellow", 
  Alistipes = "deeppink1",
  Aureibacter = "plum2", 
  Endozoicomonas = "gold", 
   Desulfobotulus= "magenta4", 
  Fibrobacter = "red1",
  Desulfovibrio="black", 
  "dgA-11 gut group" = "firebrick",
  Epulopiscium = "limegreen",
  Ferrimonas = "grey", 
  Odoribacter ="magenta",
  Rikenella = "darkslategray2",
  "Rikenellaceae RC9 gut group" = "yellow",
  Ruminococcus = "#B5D6AA",
  Romboutsia = "palevioletred1",
  Treponema = "royalblue",
  Vibrio = "darkolivegreen1",
  Other = "darkgray")
```


```{r genus-compostion}
Genus_df <-
  scaled_rooted_physeq %>%
  # agglomerate at the phylum level
  tax_glom(taxrank = "Genus") %>%
  # Transform counts to relative abundance
  transform_sample_counts(function (x) {x/sum(x)}) %>%
  # Melt to a long format
  psmelt() %>%
  # Filter out phyla that are less than one percent - get rid of low abundant Phyla
  dplyr::filter(Abundance > 0.01)
  
# Facet by gut section
Genus_df %>%
  # Warning: Its important to have one sample per x value,
  # Otherwise, it will take the sum between multiple samples
  ggplot(aes(x = names, y = Abundance, fill = Genus)) +
  facet_grid(.~gut_section) + 
  geom_bar(stat = "identity", color = "black") +
  scale_fill_manual(values = Genus_colors) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1))

level_order <- 
  c('568_4','E1_4','E2_4','E3_4','568_5', '581_5','611_5', 'E05_5', 
    'E1_5', 'E2_5', 'E3_5') 

Genus_df %>%
  # Warning: Its important to have one sample per x value,
  # Otherwise, it will take the sum between multiple samples
  ggplot(aes(x = factor(names, level = level_order), 
             y = Abundance, fill = Genus)) +
  #facet_grid(.~gut_section) + 
  geom_bar(stat = "identity", color = "black") +
  scale_fill_manual(values = Genus_colors) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1)) +
  labs(y ="Relative Abundance", x = "Sample")

# Bacteroidota
Genus_df %>%
  dplyr::filter(Phylum == "Bacteroidota") %>%
  # build the plot
  ggplot(aes(x = gut_section, y = Abundance,
             fill = gut_section, color = gut_section)) + 
  facet_wrap(.~Genus, scales = "free_y", nrow = 1) +
  geom_boxplot(alpha = 0.5, outlier.shape = NA) +
  # outliers not plotted here in boxplot
  geom_jitter() +
  theme_bw() +
  labs(title = "Bacteroidota genus Abundance") +
  scale_color_manual(values = gut_section_colors) +
  scale_fill_manual(values = gut_section_colors) +
   stat_compare_means()

# Bacillota
Genus_df %>%
  dplyr::filter(Phylum == "Bacillota") %>%
  # build the plot
  ggplot(aes(x = gut_section, y = Abundance,
             fill = gut_section, color = gut_section)) + 
  facet_wrap(.~Genus, scales = "free_y", nrow = 1) +
  geom_boxplot(alpha = 0.5, outlier.shape = NA) +
  # outliers not plotted here in boxplot
  geom_jitter() +
  theme_bw() +
  labs(title = "Bacillota Genus Abundance") +
  scale_color_manual(values = gut_section_colors) +
  scale_fill_manual(values = gut_section_colors) 

# Pseudomonadota
Genus_df %>%
  dplyr::filter(Phylum == "Pseudomonadota") %>%
  # build the plot
  ggplot(aes(x = gut_section, y = Abundance,
             fill = gut_section, color = gut_section)) + 
  facet_wrap(.~Genus, scales = "free_y", nrow = 1) +
  geom_boxplot(alpha = 0.5, outlier.shape = NA) +
  # outliers not plotted here in boxplot
  geom_jitter() +
  theme_bw() +
  labs(title = "Pseudomonadota Genus Abundance") +
  scale_color_manual(values = gut_section_colors) +
  scale_fill_manual(values = gut_section_colors)

```

```{r Genus-stats}
# Akkermansia
Genus_df %>%
  dplyr::filter(Genus == "Akkermansia") %>%
  # build the plot
  ggplot(aes(x = gut_section, y = Abundance,
             fill = gut_section, color = gut_section)) + 
  facet_wrap(.~Genus, scales = "free_y", nrow = 1) +
  geom_boxplot(alpha = 0.5, outlier.shape = NA) +
  # outliers not plotted here in boxplot
  geom_jitter() +
  theme_bw() +
  labs(title = "Akkermansia Genus Abundance") +
  scale_color_manual(values = gut_section_colors) +
  scale_fill_manual(values = gut_section_colors) #+
  #stat_compare_means(paired = FALSE)


# Desulfovibrio
Genus_df %>%
  dplyr::filter(Genus == "Desulfovibrio") %>%
  # build the plot
  ggplot(aes(x = gut_section, y = Abundance,
             fill = gut_section, color = gut_section)) + 
  facet_wrap(.~Genus, scales = "free_y", nrow = 1) +
  geom_boxplot(alpha = 0.5, outlier.shape = NA) +
  # outliers not plotted here in boxplot
  geom_jitter() +
  theme_bw() +
  labs(title = "Desulfovibrio Genus Abundance") +
  scale_color_manual(values = gut_section_colors) +
  scale_fill_manual(values = gut_section_colors) +
  stat_compare_means()


# dg A -11 gut group
Genus_df %>%
  dplyr::filter(Genus == "dgA-11 gut group") %>%
  # build the plot
  ggplot(aes(x = gut_section, y = Abundance,
             fill = gut_section, color = gut_section)) + 
  facet_wrap(.~Genus, scales = "free_y", nrow = 1) +
  geom_boxplot(alpha = 0.5, outlier.shape = NA) +
  # outliers not plotted here in boxplot
  geom_jitter() +
  theme_bw() +
  labs(title = "dgA- 11 gut group Genus Abundance") +
  scale_color_manual(values = gut_section_colors) +
  scale_fill_manual(values = gut_section_colors) 

#Fibrobacter
Genus_df %>%
  dplyr::filter(Genus == "Fibrobacter") %>%
  # build the plot
  ggplot(aes(x = gut_section, y = Abundance,
             fill = gut_section, color = gut_section)) + 
  facet_wrap(.~Genus, scales = "free_y", nrow = 1) +
  geom_boxplot(alpha = 0.5, outlier.shape = NA) +
  # outliers not plotted here in boxplot
  geom_jitter() +
  theme_bw() +
  labs(title = "Fibrobacter Genus Abundance") +
  scale_color_manual(values = gut_section_colors) +
  scale_fill_manual(values = gut_section_colors) 

#Treponema
Genus_df %>%
  dplyr::filter(Genus == "Treponema") %>%
  # build the plot
  ggplot(aes(x = gut_section, y = Abundance,
             fill = gut_section, color = gut_section)) + 
  facet_wrap(.~Genus, scales = "free_y", nrow = 1) +
  geom_boxplot(alpha = 0.5, outlier.shape = NA) +
  # outliers not plotted here in boxplot
  geom_jitter() +
  theme_bw() +
  labs(title = "Treponema Genus Abundance") +
  scale_color_manual(values = gut_section_colors) +
  scale_fill_manual(values = gut_section_colors) 
```


# Downloading files
```{r downloading-files}
# Downloading otu_table and asv_otu tables for Esther

write.csv(scaled_rooted_physeq@otu_table,
          "data/05_community_analysis/otu_table.csv")
write.csv(scaled_rooted_physeq@tax_table, 
          "data/05_community_analysis/tax_table.csv")
```


# Session Information
# For Reproducibility
```{r session-info}
#Ensure reproducibility
devtools::session_info()
```
